+++ src/ext_widgets/scroll_view/list_view.c
@@ -73,7 +73,7 @@ static ret_t list_view_on_event(widget_t* widget, event_t* e) {
   switch (e->type) {
     case EVT_WHEEL: {
       wheel_event_t* evt = (wheel_event_t*)e;
-      int32_t delta = evt->dy;
+      int32_t delta = -evt->dy;
       if (list_view->scroll_bar != NULL) {
         scroll_bar_add_delta(list_view->scroll_bar, delta);
         log_debug("wheel: %d\n", delta);