+++ tools/font_gen/font_gen.c
@@ -40,11 +40,9 @@ static int char_cmp(const void* a, const void* b) {
 }
 
 ret_t font_gen(font_t* font, uint16_t font_size, const char* str, const char* output_filename) {
-  char name[TK_NAME_LEN + 1];
   uint8_t* buff = (uint8_t*)TKMEM_ALLOC(MAX_BUFF_SIZE);
   uint32_t size = font_gen_buff(font, font_size, str, buff, MAX_BUFF_SIZE);
 
-  filename_to_name(output_filename, name, sizeof(name));
   output_res_c_source(output_filename, ASSET_TYPE_FONT, ASSET_TYPE_FONT_BMP, buff, size);
 
   TKMEM_FREE(buff);